+++ autodispose-android-archcomponents/src/main/java/com/uber/autodispose/android/archcomponents/LifecycleOwnerScopeProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.uber.autodispose.android;
+package com.uber.autodispose.android.archcomponents;
 
 import android.arch.lifecycle.Lifecycle;
 import android.arch.lifecycle.LifecycleOwner;
